+++ checkout-ui-custom/src/_js/_countries.js
@@ -490,6 +490,13 @@ module.exports._countriesrules = {
     number: false,
     postalCode: true,
   },
+  PRI: {
+    state: 'administrative_area_level_1',
+    city: 'administrative_are_level_2',
+    neighborhood: 'locality',
+    number: false,
+    postalCode: true,
+  },
 }
 
 // _countries.find(i=>i[0]=="US")
@@ -4456,15 +4463,15 @@ module.exports._cities = [
     regions: [
       {
         name: 'Bjelovarsko-bilogorska županija',
-        shortCode: 'Bjelovar-Bilogora County',
+        shortCode: 'Bjelovarsko-bilogorska županija',
       },
       {
         name: 'Brodsko-posavska županija',
-        shortCode: 'Brod-Posavina County',
+        shortCode: 'Brodsko-posavska županija',
       },
       {
         name: 'Dubrovačko-neretvanska županija',
-        shortCode: 'Dubrovnik-Neretva County',
+        shortCode: 'Dubrovačko-neretvanska županija',
       },
       {
         name: 'Grad Zagreb',
@@ -4472,27 +4479,27 @@ module.exports._cities = [
       },
       {
         name: 'Istarska županija',
-        shortCode: 'Istria County',
+        shortCode: 'Istarska županija',
       },
       {
         name: 'Karlovačka županija',
-        shortCode: 'Karlovac County',
+        shortCode: 'Karlovačka županija',
       },
       {
         name: 'Koprivničko-križevačka županija',
-        shortCode: 'Koprivnica-Križevci County',
+        shortCode: 'Koprivničko-križevačka županija',
       },
       {
         name: 'Krapinsko-zagorska županija',
-        shortCode: 'Krapina-Zagorje County',
+        shortCode: 'Krapinsko-zagorska županija',
       },
       {
         name: 'Ličko-senjska županija',
-        shortCode: 'Lika-Senj County',
+        shortCode: 'Ličko-senjska županija',
       },
       {
         name: 'Međimurska županija',
-        shortCode: 'Međimurje County',
+        shortCode: 'Međimurska županija',
       },
       {
         name: 'Osječko-baranjska županija',
@@ -4500,43 +4507,43 @@ module.exports._cities = [
       },
       {
         name: 'Požeško-slavonska županija',
-        shortCode: 'Požega-Slavonia County',
+        shortCode: 'Požeško-slavonska županija',
       },
       {
         name: 'Primorsko-goranska županija',
-        shortCode: 'Primorje-Gorski Kotar County',
+        shortCode: 'Primorsko-goranska županija',
       },
       {
         name: 'Sisačko-moslavačka županija',
-        shortCode: 'Sisak-Moslavina County',
+        shortCode: 'Sisačko-moslavačka županija',
       },
       {
         name: 'Splitsko-dalmatinska županija',
-        shortCode: 'Split-Dalmatia County',
+        shortCode: 'Splitsko-dalmatinska županija',
       },
       {
         name: 'Šibensko-kninska županija',
-        shortCode: 'Šibenik-Knin County',
+        shortCode: 'Šibensko-kninska županija',
       },
       {
         name: 'Varaždinska županija',
-        shortCode: 'Varaždin County',
+        shortCode: 'Varaždinska županija',
       },
       {
         name: 'Virovitičko-podravska županija',
-        shortCode: 'Virovitica-Podravina County',
+        shortCode: 'Virovitičko-podravska županija',
       },
       {
         name: 'Vukovarsko-srijemska županija',
-        shortCode: 'Vukovar-Srijem County',
+        shortCode: 'Vukovarsko-srijemska županija',
       },
       {
         name: 'Zadarska županija',
-        shortCode: 'Zadar County',
+        shortCode: 'Zadarska županija',
       },
       {
         name: 'Zagrebačka županija',
-        shortCode: 'Zagreb County',
+        shortCode: 'Zagrebačka županija',
       },
     ],
   },
@@ -4656,59 +4663,59 @@ module.exports._cities = [
     regions: [
       {
         name: 'Hlavní město Praha',
-        shortCode: 'PR',
+        shortCode: 'Hlavní město Praha',
       },
       {
         name: 'Jihočeský kraj',
-        shortCode: 'JC',
+        shortCode: 'Jihočeský kraj',
       },
       {
         name: 'Jihomoravský kraj',
-        shortCode: 'JM',
+        shortCode: 'Jihomoravský kraj',
       },
       {
         name: 'Karlovarský kraj',
-        shortCode: 'KA',
+        shortCode: 'Karlovarský kraj',
       },
       {
         name: 'Královéhradecký kraj',
-        shortCode: 'KR',
+        shortCode: 'Královéhradecký kraj',
       },
       {
         name: 'Liberecký kraj',
-        shortCode: 'LI',
+        shortCode: 'Liberecký kraj',
       },
       {
         name: 'Moravskoslezský kraj',
-        shortCode: 'MO',
+        shortCode: 'Moravskoslezský kraj',
       },
       {
         name: 'Olomoucký kraj',
-        shortCode: 'OL',
+        shortCode: 'Olomoucký kraj',
       },
       {
         name: 'Pardubický kraj',
-        shortCode: 'PA',
+        shortCode: 'Pardubický kraj',
       },
       {
         name: 'Plzeňský kraj',
-        shortCode: 'PL',
+        shortCode: 'Plzeňský kraj',
       },
       {
         name: 'Středočeský kraj',
-        shortCode: 'ST',
+        shortCode: 'Středočeský kraj',
       },
       {
         name: 'Ústecký kraj',
-        shortCode: 'US',
+        shortCode: 'Ústecký kraj',
       },
       {
         name: 'Vysočina',
-        shortCode: 'VY',
+        shortCode: 'Kraj Vysočina',
       },
       {
         name: 'Zlínský kraj',
-        shortCode: 'ZL',
+        shortCode: 'Zlínský kraj',
       },
     ],
   },
@@ -13427,222 +13434,295 @@ module.exports._cities = [
     regions: [
       {
         name: 'Adjuntas',
+        shortCode: 'Adjuntas',
       },
       {
         name: 'Aguada',
+        shortCode: 'Aguada',
       },
       {
         name: 'Aguadilla',
+        shortCode: 'Aguadilla',
       },
       {
         name: 'Aguas Buenas',
+        shortCode: 'Aguas Buenas',
       },
       {
         name: 'Aibonito',
+        shortCode: 'Aibonito',
       },
       {
         name: 'Anasco',
+        shortCode: 'Anasco',
       },
       {
         name: 'Arecibo',
+        shortCode: 'Arecibo',
       },
       {
         name: 'Arroyo',
+        shortCode: 'Arroyo',
       },
       {
         name: 'Barceloneta',
+        shortCode: 'Barceloneta',
       },
       {
         name: 'Barranquitas',
+        shortCode: 'Barranquitas',
       },
       {
         name: 'Bayamon',
+        shortCode: 'Bayamon',
       },
       {
         name: 'Cabo Rojo',
+        shortCode: 'Cabo Rojo',
       },
       {
         name: 'Caguas',
+        shortCode: 'Caguas',
       },
       {
         name: 'Camuy',
+        shortCode: 'Camuy',
       },
       {
         name: 'Canovanas',
+        shortCode: 'Canovanas',
       },
       {
         name: 'Carolina',
+        shortCode: 'Carolina',
       },
       {
         name: 'Cat',
+        shortCode: 'Cat',
       },
       {
         name: 'Ceiba',
+        shortCode: 'Ceiba',
       },
       {
         name: 'Ciales',
+        shortCode: 'Ciales',
       },
       {
         name: 'Cidra',
+        shortCode: 'Cidra',
       },
       {
         name: 'Coamo',
+        shortCode: 'Coamo',
       },
       {
         name: 'Comerio',
+        shortCode: 'Comerio',
       },
       {
         name: 'Corozal',
+        shortCode: 'Corozal',
       },
       {
         name: 'Culebra',
+        shortCode: 'Culebra',
       },
       {
         name: 'Dorado',
+        shortCode: 'Dorado',
       },
       {
         name: 'Fajardo',
+        shortCode: 'Fajardo',
       },
       {
         name: 'Florida',
+        shortCode: 'Florida',
       },
       {
         name: 'Guanica',
+        shortCode: 'Guanica',
       },
       {
         name: 'Guayama',
+        shortCode: 'Guayama',
       },
       {
         name: 'Guayanilla',
+        shortCode: 'Guayanilla',
       },
       {
         name: 'Guaynabo',
+        shortCode: 'Guaynabo',
       },
       {
         name: 'Gurabo',
+        shortCode: 'Gurabo',
       },
       {
         name: 'Hatillo',
+        shortCode: 'Hatillo',
       },
       {
         name: 'Hormigueros',
+        shortCode: 'Hormigueros',
       },
       {
         name: 'Humacao',
+        shortCode: 'Humacao',
       },
       {
         name: 'Isabe',
+        shortCode: 'Isabe',
       },
       {
         name: 'Juana Diaz',
+        shortCode: 'Juana Diaz',
       },
       {
         name: 'Juncos',
+        shortCode: 'Juncos',
       },
       {
         name: 'Lajas',
+        shortCode: 'Lajas',
       },
       {
         name: 'Lares',
+        shortCode: 'Lares',
       },
       {
         name: 'Las Marias',
+        shortCode: 'Las Marias',
       },
       {
         name: 'Las oiza',
+        shortCode: 'Las oiza',
       },
       {
         name: 'Luquillo',
+        shortCode: 'Luquillo',
       },
       {
         name: 'Manati',
+        shortCode: 'Manati',
       },
       {
         name: 'Maricao',
+        shortCode: 'Maricao',
       },
       {
         name: 'Maunabo',
+        shortCode: 'Maunabo',
       },
       {
         name: 'Mayaguez',
+        shortCode: 'Mayaguez',
       },
       {
         name: 'Moca',
+        shortCode: 'Moca',
       },
       {
         name: 'Morovis',
+        shortCode: 'Morovis',
       },
       {
         name: 'Naguabo',
+        shortCode: 'Naguabo',
       },
       {
         name: 'Naranjito',
+        shortCode: 'Naranjito',
       },
       {
         name: 'Orocovis',
+        shortCode: 'Orocovis',
       },
       {
         name: 'Patillas',
+        shortCode: 'Patillas',
       },
       {
         name: 'Penuelas',
+        shortCode: 'Penuelas',
       },
       {
         name: 'Ponce',
+        shortCode: 'Ponce',
       },
       {
         name: 'Quebradillas',
+        shortCode: 'Quebradillas',
       },
       {
         name: 'Rincon',
+        shortCode: 'Rincon',
       },
       {
         name: 'Rio Grande',
+        shortCode: 'Rio Grande',
       },
       {
         name: 'Sabana linas',
+        shortCode: 'Sabana linas',
       },
       {
         name: 'San German',
+        shortCode: 'San German',
       },
       {
         name: 'San Juan',
+        shortCode: 'San Juan',
       },
       {
         name: 'San Lorenzo',
+        shortCode: 'San Lorenzo',
       },
       {
         name: 'San Sebastian',
+        shortCode: 'San Sebastian',
       },
       {
         name: 'Santa Isabel',
+        shortCode: 'Santa Isabel',
       },
       {
         name: 'Toa Alta',
+        shortCode: 'Toa Alta',
       },
       {
         name: 'Toa Baja',
+        shortCode: 'Toa Baja',
       },
       {
         name: 'Trujillo Alto',
+        shortCode: 'Trujillo Alto',
       },
       {
         name: 'Utuado',
+        shortCode: 'Utuado',
       },
       {
         name: 'Vega Alta',
+        shortCode: 'Vega Alta',
       },
       {
         name: 'Vega ues',
+        shortCode: 'Vega ues',
       },
       {
         name: 'Villalba',
+        shortCode: 'Villalba',
       },
       {
         name: 'Yabucoa',
+        shortCode: 'Yabucoa',
       },
       {
         name: 'Yauco',
+        shortCode: 'Yauco',
       },
     ],
   },